+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - changed carbon unit to KGCO2E in myems-api
 - updated data source page in myems-admin
 - changed login page in myems-admin
+- changed carbon dioxide emission factor from fixed value to timeofuse values in database, myems-api, myems-admin and myems-aggregation
 ### Fixed
 - fixed issue of meter realtime report in myems-web
 - fixed issue of space energy item in myems-web